requirements.txt
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.19.5
Pillow>=8.2.0
matplotlib>=3.4.2
scikit-learn>=0.24.2
tqdm>=4.61.1
PyQt5>=5.15.4
segmentation-models-pytorch>=0.2.1
albumentations>=1.0.3
pandas>=1.3.0
opencv-python>=4.5.3
timm>=0.4.12